requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yadisk==3.2.0
yarl==1.18.3
